packages/vlfs_core/pyproject.toml
[project]
name = "vlfs-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Federico Vela", email = "fedevela@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.72.0",
    "pathspec>=1.0.4",
    "pyyaml>=6.0.3",
    "sqlite-vec>=0.1.9",
]

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"


